+++ adhocracy-plus/config/settings/production.py
@@ -35,6 +35,11 @@ except NameError:
     pass
 
 CELERY_BEAT_SCHEDULE = {
+    "publish-results-reminders-hourly": {
+        "task": "send_publish_results_reminders",
+        "schedule": timedelta(hours=1),
+        "args": (),
+    },
     "send-recently-started-project-notifications": {
         "task": "send_recently_started_project_notifications",
         "schedule": timedelta(days=3),